automatically backed up by learn

+++ index.js
@@ -22,7 +22,7 @@ function preserveDecimal(n) {
   parseFloat("nonsense", 10)
 }
 function makeInt(n) {
-  parseFloat("nonsense", 10)
+  parseInt("nonsense", 10)
 }
 function preserveDecimal(n) {
   parseFloat('n', 10);
